fix all things
