+++ tests/conftest.py
@@ -1,6 +1,8 @@
 """Pytest utility classes, functions and fixtures."""
 
+import inspect
 import os
+import types
 
 import numpy as np
 import pytest
@@ -92,6 +94,119 @@ def pytorch_error_msg(a, b, rtol, atol):
             diff = torch.abs(a - b)
             msg = msg + f"\ndifference \n{diff}\nrtol {rtol}\natol {atol}"
     return msg
+
+
+def copy_func(f, name=None):
+    """
+    Return a function with same code, globals, defaults, closure, and
+    name (or provide a new name)
+    """
+    fn = types.FunctionType(
+        f.__code__, f.__globals__, name or f.__name__, f.__defaults__, f.__closure__
+    )
+    fn.__dict__.update(f.__dict__)
+    return fn
+
+
+class Parametrizer(type):
+    """Metaclass for test classes.
+
+    Note: A test class is a class that inherits from TestCase.
+    For example, `class TestEuclidean(TestCase)` defines
+    a test class.
+
+    The Parametrizer helps its test class by pairing:
+    - the different test functions of the test class:
+      - e.g. the test function `test_belongs`,
+    - with different test data, generated by auxiliary test data functions
+      - e.g. the test data function `belongs_data` that generates data
+      to test the function `belongs`.
+
+    As such, Parametrizer acts as a "metaclass" of its test class:
+    `class TestEuclidean(TestCase, metaclass=Parametrizer)`.
+
+    Specifically, Parametrizer decorates every test function inside
+    its test class with pytest.mark.parametrizer, with the exception
+    of the test class' class methods and static methods.
+
+    Two conventions need to be respected:
+    1. The test class should contain an attribute named 'testing_data'.
+      - `testing_data` is an object inheriting from `TestData`.
+    2. Every test function should have its corresponding test data function created
+    inside the TestData object called `testing_data`.
+
+    A sample test class looks like this:
+
+    ```
+    class TestEuclidean(TestCase, metaclass=Parametrizer):
+        class TestDataEuclidean(TestData):
+            def belongs_data():
+                ...
+                return self.generate_tests(...)
+        testing_data = TestDataEuclidean()
+        def test_belongs():
+            ...
+    ```
+    Parameters
+    ----------
+    cls : child class of TestCase
+        Test class, i.e. a class inheriting from TestCase
+    name : str
+        Name of the test class
+    bases : TestCase
+        Parent class of the test class: TestCase.
+    attrs : dict
+        Attributes of the test class, for example its methods,
+        stored in a dictionnary as (key, value) when key gives the
+        name of the attribute (for example the name of the method),
+        and value gives the actual attribute (for example the method
+        itself.)
+
+    References
+    ----------
+    More on pytest's parametrizers can be found here:
+    https://docs.pytest.org/en/6.2.x/parametrize.html
+    """
+
+    def __new__(cls, name, bases, attrs):
+        """Decorate the test class' methods with pytest."""
+        for base in bases:
+            test_fn_list = [fn for fn in dir(base) if fn.startswith("test") is True]
+            for test_fn in test_fn_list:
+                attrs[test_fn] = copy_func(getattr(base, test_fn))
+
+        for attr_name, attr_value in attrs.copy().items():
+            if isinstance(attr_value, types.FunctionType):
+
+                if ("skip_" + attr_name, True) not in locals()["attrs"].items():
+                    args_str = ", ".join(inspect.getfullargspec(attr_value)[0][1:])
+                    data_fn_str = attr_name[5:] + "_test_data"
+                    if "testing_data" not in locals()["attrs"]:
+                        raise Exception(
+                            "Testing class class doesn't have class attribute"
+                            " named 'testing_data'"
+                        )
+                    if not hasattr(locals()["attrs"]["testing_data"], data_fn_str):
+                        raise Exception(
+                            "testing_data object doesn't have '{}' function for "
+                            "pairing with '{}'".format(data_fn_str, attr_name)
+                        )
+                    test_data = getattr(
+                        locals()["attrs"]["testing_data"], data_fn_str
+                    )()
+                    if test_data is None:
+                        raise Exception(
+                            "'{}' returned None. should be list".format(data_fn_str)
+                        )
+
+                    attrs[attr_name] = pytest.mark.parametrize(
+                        args_str,
+                        test_data,
+                    )(attr_value)
+                else:
+                    attrs[attr_name] = pytest.mark.skip("skipped")(attr_value)
+
+        return super(Parametrizer, cls).__new__(cls, name, bases, attrs)
 
 
 class TestCase:
