+++ tests/tests_geomstats/test_lie_algebra.py
@@ -6,9 +6,8 @@ from geomstats.geometry.invariant_metric import InvariantMetric
 from geomstats.geometry.skew_symmetric_matrices import SkewSymmetricMatrices
 from geomstats.geometry.special_euclidean import SpecialEuclidean
 from geomstats.geometry.special_orthogonal import SpecialOrthogonal
-from tests.conftest import TestCase
+from tests.conftest import Parametrizer, TestCase
 from tests.data_generation import TestData
-from tests.parametrizers import Parametrizer
 
 
 class TestLieAlgebra(TestCase, metaclass=Parametrizer):
